demux:asf: log the seconds as integers
As it was before

+++ modules/demux/asf/asf.c
@@ -356,7 +356,7 @@ static int SeekIndex( demux_t *p_demux, vlc_tick_t i_date, float f_pos )
     asf_object_index_t *p_index;
 
     msg_Dbg( p_demux, "seek with index: %i seconds, position %f",
-             i_date >= 0 ? SEC_FROM_VLC_TICK(i_date) : -1, f_pos );
+             i_date >= 0 ? (int)SEC_FROM_VLC_TICK(i_date) : -1, f_pos );
 
     if( i_date < 0 )
         i_date = p_sys->i_length * f_pos;
